LayoutTests/inspector-obsolete/http-tests/network/network-shared-worker-expected.txt

CONSOLE MESSAGE: line 14: Done.
Tests that shared worker content is available in network panel.

Bug 65929
http://127.0.0.1:8000/inspector/network/resources/shared-worker.js
resource.type: script
resource.content before requesting content: undefined
resource.content after requesting content: onconnect = function(event)
{
    var port = event.ports[0];
    port.postMessage("Done.");
}
